refactor(theme-tool): refactor theme-tool done
