--- packages/theme/old-theme-index.js
@@ -1,6 +1,0 @@
-export default {
-  id: 'tiny-old-theme',
-  name: 'OldTheme',
-  cnName: '旧的主题',
-  data: `#CSS#`
-}
