+++ packages/theme/src/button/old-theme.less
@@ -3,5 +3,5 @@
 
 .@{button-prefix-cls} {
   // 默认时按钮字重
-  --tv-Button-font-weight: var(--tv-font-weight-thin);
+  --tv-Button-border-radius: 6px;
 }
